requirements.txt
flask
face_recognition
opencv-python
requests
python-telegram-bot
pillow